stock api changes

+++ app/src/main/java/stock/com/ui/dashboard/my_contest/adapter/FinishedAdapter.kt
@@ -74,7 +74,7 @@ class FinishedAdapter(
             mContext.startActivity(
                 Intent(mContext, LiveScoreActivity::class.java)
                     .putExtra(StockConstant.CONTESTID, contest.get(position).contestId)
-                    .putExtra(StockConstant.MARKETID, contest.get(position).mid.toInt())
+                    .putExtra(StockConstant.MARKETID, contest.get(position).mid)
             )
         }
     }
